+++ src/pages/CommunityPage/CommunityCreatePage.tsx
@@ -1,14 +1,17 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useMutation } from 'react-query';
 import { Helmet } from 'react-helmet-async';
-import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
 import styled from 'styled-components';
 
 import { supabase } from '@/client';
 import { useAuthStore } from '@store/useAuthStore';
 import useCreateStore from '@store/useCreateStore';
+import CommunityCreateDetail from '@/components/CommunityPage/CommunityCreateDetail';
+import CommunityCreateIntroduction from '@/components/CommunityPage/CommunityCreateIntroduction';
+import CommunityCreateAuthorInfo from '@components/CommunityPage/CommunityCreateAuthorInfo';
+import CommunityCreateDeadline from '@components/CommunityPage/CommunityCreateDeadline';
 
 interface CreateData {
   title: string;
@@ -16,15 +19,12 @@ interface CreateData {
   user_id: string | number;
 }
 
-
-
-
 function CommunityCreatePage() {
-  const [title, setTitle] = useState('');
-  const [contents, setContent] = useState('');
   const user = useAuthStore((state) => state.user);
-  const [userEmail, setUserEmail] = useState<string>('');
+  const [userEmail] = useState<string>('');
   const { addComment } = useCreateStore();
+  const [title, setTitle] = useState('');
+  const [contents, setContents] = useState('');
   const [people, setPeople] = useState('');
   const [division, setDivision] = useState('');
   const [progress, setProgress] = useState('');
@@ -37,30 +37,7 @@ function CommunityCreatePage() {
   const handleTitleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setTitle(e.target.value);
   };
-
-  const handleContentChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setContent(e.target.value);
-  };
-const handlePeopleChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-  setPeople(e.target.value);
-};
-const handleDivisionChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-  setDivision(e.target.value);
  
-};
-const handleProgressChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-  setProgress(e.target.value);
-};
-const handleTag1Change = (e: React.ChangeEvent<HTMLSelectElement>) => {
-  setStack1(e.target.value);
-};
-const handleTag2Change = (e: React.ChangeEvent<HTMLSelectElement>) => {
-  setStack2(e.target.value);
-};
-const handleTag3Change = (e: React.ChangeEvent<HTMLSelectElement>) => {
-  setStack3(e.target.value);
-};
-
   const createPost = useMutation(async (data: CreateData) => {
     const currentDate = new Date().toISOString();
     const targetTable =
@@ -86,47 +63,11 @@ const handleTag3Change = (e: React.ChangeEvent<HTMLSelectElement>) => {
         throw new Error('Error creating post');
       }
       
-      
     } catch (error) {
       console.error('Error creating post:', error);
     }
      },
-     {
-       onSuccess: () => {
-         
-         console.log('Successfully created post!');
-        
-      },
-      onError: () => {
-        
-        console.error('Failed to create post');
-      
-      },
-      
-    });
-   
-    
-    useEffect(() => {
-    const fetchUserEmail = async () => {
-      if (user) {
-        const { data, error } = await supabase
-        .from('users')
-          .select('email')
-          .eq('id', user)
-          .single();
-          
-          if (error) {
-            console.error('Error fetching user data:', error);
-            return;
-          }
-          
-          if (data) {
-            data.email && setUserEmail(data.email);
-          }
-        }
-      };
-      fetchUserEmail();
-    }, [user]);
+    );
     
     const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
       e.preventDefault();
@@ -155,7 +96,6 @@ const handleTag3Change = (e: React.ChangeEvent<HTMLSelectElement>) => {
         <title>CommunityCreate - JUNGLE</title>
       </Helmet>
       <section>
-        <h1>CommunityCreatePage</h1>
         <div>
           <FormContainer>
             <fieldset>
@@ -173,134 +113,35 @@ const handleTag3Change = (e: React.ChangeEvent<HTMLSelectElement>) => {
                 <ul>
                   <div>
                     <Info>✏프로젝트 기본 정보를 입력해주세요.</Info>
-                    <Firstwrapper>
-                      <Li>
-                        <label>작성자</label>
-                        <div> {userEmail} </div>
-                      </Li>
-                      <Li>
-                        <label>작성일자</label>
-                        <div>{new Date().toISOString().slice(0, 10)} </div>
-                      </Li>
-                    </Firstwrapper>
-                    <Secondwrapper>
-                      <Li>
-                          <p>구분</p>
-                          <Select
-                            defaultValue="프로젝트/스터디"
-                            onChange={handleDivisionChange}
-                          >
-                            <option disabled>프로젝트/스터디</option>
-                            <option>프로젝트</option>
-                            <option>스터디</option>
-                          </Select>
-                      </Li>
-                      <Li>
-                        <p>모집인원</p>
-                        <Select
-                          defaultValue="모집인원"
-                          onChange={handlePeopleChange}
-                        >
-                          <option disabled>모집인원</option>
-                          <option>1 명</option>
-                          <option>2 명</option>
-                          <option>3 명</option>
-                          <option>4 명</option>
-                          <option>5 명</option>
-                          <option>인원 수 제한없음</option>
-                        </Select>
-                      </Li>
-                    </Secondwrapper>
-                    <Thirdwrapper>
-                      <Li>
-                        <p>진행방식</p>
-                        <Select
-                          defaultValue="진행방식"
-                          onChange={handleProgressChange}
-                        >
-                          <option disabled>진행방식</option>
-                          <option>온라인</option>
-                          <option>오프라인</option>
-                        </Select>
-                      </Li>
-                      <Li>
-                        <p>사용언어</p>
-                        <Select
-                          defaultValue="사용언어"
-                          onChange={handleTag1Change}
-                        >
-                          <option disabled>사용언어</option>
-                          <option value="javascript">javascript</option>
-                          <option value="react">react</option>
-                          <option value="ts">ts</option>
-                          <option value="next.js">next.js</option>
-                          <option value="vue">vue</option>
-                          <option value="svelte">svelte</option>
-                        </Select>
-                      </Li>
-                    </Thirdwrapper>
-                    <Fourthwrapper>
-                      <Li>
-                        <p>사용언어</p>
-                        <Select
-                          defaultValue="사용언어"
-                          onChange={handleTag2Change}
-                        >
-                          <option disabled>사용언어</option>
-                          <option value="javascript">javascript</option>
-                          <option value="react">react</option>
-                          <option value="ts">ts</option>
-                          <option value="next.js">next.js</option>
-                          <option value="vue">vue</option>
-                          <option value="svelte">svelte</option>
-                        </Select>
-                      </Li>
-                      <Li>
-                        <p>사용언어</p>
-                        <Select defaultValue="사용언어" 
-                        onChange={handleTag3Change}>
-                          <option disabled>
-                            사용언어
-                          </option>
-                          <option value="javascript">javascript</option>
-                          <option value="react">react</option>
-                          <option value="ts">ts</option>
-                          <option value="next.js">next.js</option>
-                          <option value="vue">vue</option>
-                          <option value="svelte">svelte</option>
-                        </Select>
-                      </Li>
-                    </Fourthwrapper>
+                    <CommunityCreateAuthorInfo userEmail={userEmail} />
+                    <CommunityCreateDetail
+                      division={division}
+                      people={people}
+                      progress={progress}
+                      tag1={tag1}
+                      tag2={tag2}
+                      tag3={tag3}
+                      setDivision={setDivision}
+                      setPeople={setPeople}
+                      setProgress={setProgress}
+                      setStack1={setStack1}
+                      setStack2={setStack2}
+                      setStack3={setStack3}
+                    />
                   </div>
                   <Li>
-                    <Datewrapper>
-                      <label>마감일 </label>
-                      <StyledDatePicker
-                        selected={deadline}
-                        onChange={(date) => setDeadline(date)}
-                        dateFormat="yyyy-MM-dd"
-                        placeholderText="Select a date"
-                      />
-                    </Datewrapper>
+                    <CommunityCreateDeadline
+                      deadline={deadline}
+                      setDeadline={setDeadline}
+                    />
                   </Li>
                   <Info>👩‍💻프로젝트에 대해 소개해주세요.</Info>
-                  <Li>
-                    <Label>제목</Label>
-                    <StyledInput
-                      type="text"
-                      value={title}
-                      onChange={handleTitleChange}
-                      placeholder="JUNGLE을 이용해 주셔서 감사합니다. 제목을 입력해 주세요."
-                    />
-                  </Li>
-                  <Li>
-                    <Label>내용</Label>
-                    <Styledtextarea
-                      value={contents}
-                      onChange={handleContentChange}
-                      placeholder="JUNGLE을 이용해 주셔서 감사합니다. 모집 내용을 입력해 주세요."
-                    />
-                  </Li>
+                  <CommunityCreateIntroduction
+                    title={title}
+                    setTitle={setTitle}
+                    contents={contents}
+                    setContents={setContents}
+                  />
                 </ul>
                 <SubmitWrapper>
                   <Submit type="submit">작성 완료</Submit>
@@ -350,54 +191,9 @@ const Srlegend = styled.legend`
   }
 `;
 
-const Styledtextarea = styled.textarea`
-  width: 100%;
-  height: 100px;
-  font-size: 1.25rem;
-`;
-
-const Firstwrapper = styled.div`
-  display: flex;
-  padding-top: 0.625rem;
-  `;
-
-const Secondwrapper = styled.div`
-display: flex;
-`;
-
-const Thirdwrapper = styled.div`
-  display: flex;
-`;
-
-const Fourthwrapper = styled.div`
-  display: flex;
-`;
-
-const Datewrapper = styled.div`
-  padding-top: 0.625rem;
-`;
-
-const Select = styled.select`
-  width: 100%;
-  height: 2.5rem;
-  text-align: center;
-`;
-
 const Info = styled.div`
   padding-bottom: 1.25rem;
   border-bottom: 1px solid #d8d8d8;
-`;
-
-const Label = styled.label`
-  display: block;
-  padding-bottom: 0.625rem;
-  padding-top: 0.625rem;
-`;
-
-const StyledDatePicker = styled(DatePicker)`
-  width: 100%;
-  height: 100%;
-  text-align: center;
 `;
 
   const Submit = styled.button`
